etl/load_db.py
"""
Database loading module for MySQL
"""

import mysql.connector
from mysql.connector import Error
from typing import List, Dict, Any, Optional
import logging

from etl.config import DB_CONFIG

logger = logging.getLogger(__name__)


def create_connection() -> Optional[mysql.connector.MySQLConnection]:
    """
    Create a database connection to MySQL database
    
    Returns:
        Connection object or None if connection fails
    """
    try:
        connection = mysql.connector.connect(**DB_CONFIG)
        if connection.is_connected():
            logger.info(f"Successfully connected to MySQL database: {DB_CONFIG['database']}")
            return connection
    except Error as e:
        logger.error(f"Error connecting to MySQL database: {e}")
        return None


def close_connection(conn: mysql.connector.MySQLConnection):
    """
    Close database connection
    
    Args:
        conn: Database connection
    """
    if conn and conn.is_connected():
        conn.close()
        logger.info("MySQL connection closed")


def load_transactions(transactions: List[Dict[str, Any]]):
    """
    Load multiple transactions into the MySQL database
    
    Args:
        transactions: List of transaction dictionaries
    """
    # TODO: Implement loading logic with proper error handling
    # This should insert data into TRANSACTION, USER, CATEGORY tables
    # based on the schema in database/database_setup.sql
    pass
