requirements.txt
lxml
python-dateutil
mysql-connector-python
# optional
fastapi
uvicorn